+++ lib/legion/extensions/ollama.rb
@@ -2,6 +2,8 @@
 
 require 'legion/extensions/ollama/version'
 require 'legion/extensions/ollama/helpers/client'
+require 'legion/extensions/ollama/helpers/errors'
+require 'legion/extensions/ollama/helpers/usage'
 require 'legion/extensions/ollama/runners/completions'
 require 'legion/extensions/ollama/runners/chat'
 require 'legion/extensions/ollama/runners/models'